add a new `directories` file
