+++ src/source/mod.rs
@@ -6,7 +6,7 @@ use std::{
 };
 
 use crate::{
-    metadata::{Directories, Output},
+    Output,
     recipe::parser::{GitRev, GitSource, Source},
     source::{
         checksum::Checksum,
@@ -14,6 +14,7 @@ use crate::{
     },
     system_tools::ToolError,
     tool_configuration,
+    types::Directories,
 };
 
 use fs_err as fs;
